Simplify files organization
