+++ .rubocop_todo.yml
@@ -1,16 +1,21 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-03-11 15:59:15 +0100 using RuboCop version 0.79.0.
+# on 2020-05-24 23:52:41 +0200 using RuboCop version 0.84.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 21
+# Offense count: 19
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
-  Max: 343
+  Max: 434
+
+# Offense count: 1
+# Configuration parameters: CountComments.
+Metrics/ClassLength:
+  Max: 160
 
 # Offense count: 1
 # Configuration parameters: CountComments, ExcludedMethods.
@@ -20,15 +25,8 @@ Metrics/MethodLength:
 # Offense count: 1
 Naming/AccessorMethodName:
   Exclude:
-    - 'lib/meilisearch/index/updates.rb'
+    - 'lib/meilisearch/index.rb'
 
-# Offense count: 17
+# Offense count: 5
 Style/Documentation:
   Enabled: false
-
-# Offense count: 73
-# Cop supports --auto-correct.
-# Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
-# URISchemes: http, https
-Layout/LineLength:
-  Max: 107
