--- lib/meilisearch/client/health.rb
@@ -1,22 +1,0 @@
-# frozen_string_literal: true
-
-module MeiliSearch
-  class Client < HTTPRequest
-    module Health
-      def healthy?
-        http_get '/health'
-        true
-      rescue StandardError
-        false
-      end
-
-      def health
-        http_get '/health'
-      end
-
-      def update_health(bool)
-        http_put '/health', health: bool
-      end
-    end
-  end
-end
