+++ spec/meilisearch/index/stats_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe MeiliSearch::Index::Stats do
+RSpec.describe MeiliSearch::Index do
   before(:all) do
     @documents = [
       { objectId: 123,  title: 'Pride and Prejudice' },
